{{cookiecutter.project_name}}/requirements.txt
APScheduler==3.10.4
certifi==2024.2.2
charset-normalizer==3.3.2
dominate==2.9.1
idna==3.7
mysql-connector-python==8.3.0
pytz==2024.1
requests==2.31.0
six==1.16.0
tzlocal==5.2
urllib3==2.2.1
